# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Creates a new user account. Required RegisterDto fields: fullName, email, password. Password must be at least 8 characters. This endpoint is public and does not require a token.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Registration payload. Example: { \"fullName\": \"John Doe\", \"email\": \"user@example.com\", \"password\": \"secret1234\" }.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully. Returns ServiceResultContainer<UserResponseDto>.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO validation failed."
          },
          "401": {
            "description": "Not applicable for this public endpoint."
          },
          "403": {
            "description": "Not applicable for this public endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "summary": "Register a new user account",
        "summaryHe": "יוצר חשבון חדש ומצטרף למשפחת המערכת",
        "tags": [
          "auth"
        ],
        "toolIcon": "ph-user-plus"
      }
    },
    "/auth/login": {
      "post": {
        "description": "Validates user credentials. On success, the service writes access and refresh tokens to HTTP-only cookies and returns ServiceResultContainer<UserResponseDto>. Role is numeric: 1 = Admin, 2 = User.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Login credentials. Example: { \"email\": \"user@example.com\", \"password\": \"secret1234\" }.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Tokens are set as cookies by the service.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO validation failed."
          },
          "401": {
            "description": "Invalid credentials."
          },
          "403": {
            "description": "Not applicable for this public endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "summary": "Login with email and password",
        "summaryHe": "נכנס לחשבון האישי בבטחה",
        "tags": [
          "auth"
        ],
        "toolIcon": "ph-sign-in"
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "JwtRefreshGuard reads the refresh token from the cookie or Authorization header. If valid, AuthService issues new tokens and returns ServiceResultContainer<UserResponseDto>.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New access token issued. Updated tokens are set as cookies by the service.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request."
          },
          "401": {
            "description": "Refresh token is missing, expired, revoked, or does not match the stored hash."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "summary": "Refresh access token using refresh token",
        "summaryHe": "מחדש את טוקן הגישה ברקע כדי להישאר מחובר",
        "tags": [
          "auth"
        ],
        "toolIcon": "ph-arrows-clockwise"
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Always succeeds, even when the access token is already expired: revokes the stored refresh token (best-effort, from the refresh cookie) and removes auth cookies from the response.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request."
          },
          "401": {
            "description": "Not applicable — logout never rejects."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "summary": "Logout and invalidate session",
        "summaryHe": "מתנתק מהמערכת ומסיים את סשן העבודה בבטחה",
        "tags": [
          "auth"
        ],
        "toolIcon": "ph-sign-out"
      }
    },
    "/auth/me": {
      "get": {
        "description": "Reads req.user as populated by JwtAuthGuard. No database query is made. Role is numeric: 1 = Admin, 2 = User.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "ServiceResultContainer<JwtPayload> with sub, email, numeric role, iat, and exp.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtPayloadResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request."
          },
          "401": {
            "description": "Access token is missing, expired, or invalid."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user payload",
        "summaryHe": "מציג את פרטי הפרופיל המהירים של המשתמש הנוכחי",
        "tags": [
          "auth"
        ],
        "toolIcon": "ph-user-circle"
      }
    },
...
    "/currency/current": {
      "get": {
        "description": "Retrieves current exchange rates from the external exchange-rate provider.",
        "operationId": "CurrencyController_getRates",
        "parameters": [
          {
            "name": "base",
            "required": false,
            "in": "query",
            "description": "Three-letter ISO currency code to use as the exchange-rate base. Defaults to ILS.",
            "schema": {
              "default": "ILS",
              "example": "USD",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currency rates retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyRatesResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO validation failed."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get exchange rates for a base currency",
        "summaryHe": "מציג שערי חליפין מעודכנים על בסיס מטבע נבחר",
        "tags": [
          "currency"
        ],
        "toolIcon": "ph-currency-circle-dollar"
      }
    },
    "/currency/convert": {
      "get": {
        "description": "Converts a positive amount between two supported currencies using current exchange rates.",
        "operationId": "CurrencyController_convert",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "description": "Three-letter ISO source currency code.",
            "schema": {
              "example": "USD",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Three-letter ISO target currency code.",
            "schema": {
              "example": "ILS",
              "type": "string"
            }
          },
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "description": "Positive numeric amount to convert.",
            "schema": {
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Currency amount converted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyConversionResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO validation failed."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "500": {
            "description": "Unexpected server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Convert currency amount",
        "summaryHe": "ממיר סכומי כסף בין שני מטבעות לפי השער היציג העדכני",
        "tags": [
          "currency"
        ],
        "toolIcon": "ph-currency-circle-dollar"
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "John Doe",
            "description": "Full display name for the new account. Required - must be a non-empty string. This field is not required to be unique."
          },
          "email": {
            "type": "string",
            "example": "admin@admin.com",
            "format": "email",
            "description": "Email address for the new account. Required - must be a valid email format and must be unique."
          },
          "password": {
            "type": "string",
            "example": "admin1234",
            "minLength": 8,
            "description": "Password for the new account. Minimum 8 characters. Required - stored as a bcrypt hash and never persisted as plain text."
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ]
      },
      "UserRole": {
        "type": "number",
        "enum": [
          1,
          2
        ],
        "description": "Numeric user role. 1 = Admin, 2 = User."
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique numeric user id.",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "User email address.",
            "example": "user@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Full display name of the user.",
            "example": "John Doe"
          },
          "role": {
            "description": "Numeric user role. 1 = Admin, 2 = User.",
            "example": 2,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was created.",
            "example": "2026-05-12T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was last updated.",
            "example": "2026-05-12T11:00:00Z"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user last successful login, or null if the user never logged in.",
            "example": "2026-05-20T10:00:00Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "createdAt",
          "updatedAt",
          "lastLoginAt"
        ]
      },
      "UserResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "User retrieved successfully"
          },
          "result": {
            "description": "Public user response payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@admin.com",
            "format": "email",
            "description": "Registered email address of the user. Must be a valid email format (user@domain.com). Case-insensitive - \"Admin@Admin.com\" and \"admin@admin.com\" are treated as the same."
          },
          "password": {
            "type": "string",
            "example": "admin1234",
            "minLength": 4,
            "description": "Account password. Minimum 4 characters for login validation. Sent over HTTPS; hashing and comparison are handled server-side. Returns 401 if this does not match the stored hash."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LogoutPayloadDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "True when logout completed.",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
      "LogoutResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "Logged out successfully"
          },
          "result": {
            "description": "Logout result payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LogoutPayloadDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "JwtPayloadResponseDto": {
        "type": "object",
        "properties": {
          "sub": {
            "type": "number",
            "description": "Authenticated user id from the JWT subject.",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "Authenticated user email from the JWT payload.",
            "example": "user@example.com"
          },
          "role": {
            "description": "Numeric user role from the JWT payload. 1 = Admin, 2 = User.",
            "example": 2,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "iat": {
            "type": "number",
            "description": "JWT issued-at timestamp in seconds.",
            "example": 1770000000
          },
          "exp": {
            "type": "number",
            "description": "JWT expiry timestamp in seconds.",
            "example": 1770000900
          }
        },
        "required": [
          "sub",
          "email",
          "role",
          "iat",
          "exp"
        ]
      },
      "JwtPayloadResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "Current user retrieved successfully"
          },
          "result": {
            "description": "Decoded JWT payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/JwtPayloadResponseDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
...
      "CurrencyRatesResponseDto": {
        "type": "object",
        "properties": {
          "base": {
            "type": "string",
            "description": "Base currency code used for the rates.",
            "example": "USD"
          },
          "date": {
            "type": "string",
            "description": "Exchange-rate update date from the provider.",
            "example": "2026-06-07"
          },
          "rates": {
            "type": "object",
            "description": "Map of currency code to exchange-rate value.",
            "example": {
              "ILS": 3.68,
              "EUR": 0.92,
              "GBP": 0.78
            },
            "additionalProperties": {
              "type": "number"
            }
          }
        },
        "required": [
          "base",
          "date",
          "rates"
        ]
      },
      "CurrencyRatesResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "Currency rates retrieved successfully"
          },
          "result": {
            "description": "Currency rates payload.",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CurrencyRatesResponseDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "CurrencyConversionResponseDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "Source currency code.",
            "example": "USD"
          },
          "to": {
            "type": "string",
            "description": "Target currency code.",
            "example": "ILS"
          },
          "amount": {
            "type": "number",
            "description": "Original amount requested for conversion.",
            "example": 100
          },
          "rate": {
            "type": "number",
            "description": "Exchange rate from source currency to target currency.",
            "example": 3.68
          },
          "result": {
            "type": "number",
            "description": "Converted amount.",
            "example": 368
          },
          "date": {
            "type": "string",
            "description": "Exchange-rate update date from the provider.",
            "example": "2026-06-07"
          }
        },
        "required": [
          "from",
          "to",
          "amount",
          "rate",
          "result",
          "date"
        ]
      },
      "CurrencyConversionResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable result message.",
            "example": "Currency converted successfully"
          },
          "result": {
            "description": "Currency conversion payload.",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CurrencyConversionResponseDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
...
    }
  }
}